Add comments and tests
The logic is not so straight, as I used a closed form solution. So I
added some comments. Also, I made a lot of tests "by hand" as I was
trying to understand the best solution, so I added some automated tests
to be able to refactor it with confidence.